+++ bricks/shoelace/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/shoelace@0.3.1...@next-bricks/shoelace@0.3.2) (2023-05-26)
+
+**Note:** Version bump only for package @next-bricks/shoelace
+
+
+
+
+
 ## [0.3.1](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/shoelace@0.3.0...@next-bricks/shoelace@0.3.1) (2023-05-19)
 
 **Note:** Version bump only for package @next-bricks/shoelace