+++ shared/general/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/easyops-cn/next-bricks/compare/@next-shared/general@0.1.0...@next-shared/general@0.2.0) (2023-05-26)
+
+
+### Features
+
+* support workbench-aciton-list ([933cc19](https://github.com/easyops-cn/next-bricks/commit/933cc19ae9588fbc6785653276f520ea9dc84324))
+
+
+
+
+
 # 0.1.0 (2023-05-22)